# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "additionalProperties": false,
        "properties": {
          "acknowledged_dq_codes": {
            "items": {
              "type": "string"
            },
            "title": "Acknowledged Dq Codes",
            "type": "array"
          },
          "comment": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Comment",
            "type": "string"
          },
          "mapping_confirmed": {
            "default": false,
            "title": "Mapping Confirmed",
            "type": "boolean"
          }
        },
        "required": [
          "comment"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "Body_createImport": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "portfolio_code": {
            "title": "Portfolio Code",
            "type": "string"
          },
          "portfolio_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Name"
          }
        },
        "required": [
          "file",
          "portfolio_code"
        ],
        "title": "Body_createImport",
        "type": "object"
      },
...
      "Body_uploadReferenceDictionary": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_uploadReferenceDictionary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImportComparisonResponse": {
        "additionalProperties": false,
        "properties": {
          "baseline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComparisonIdentity"
              },
              {
                "type": "null"
              }
            ]
          },
          "current": {
            "$ref": "#/components/schemas/ComparisonIdentity"
          },
          "lot_changes": {
            "$ref": "#/components/schemas/LotChanges"
          },
          "metrics": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ComparisonMetric"
            },
            "title": "Metrics",
            "type": "object"
          }
        },
        "required": [
          "current",
          "baseline",
          "metrics",
          "lot_changes"
        ],
        "title": "ImportComparisonResponse",
        "type": "object"
      },
...
      "ImportListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ImportRecord"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ImportListResponse",
        "type": "object"
      },
      "ImportRecord": {
        "additionalProperties": false,
        "properties": {
          "approved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "dq_counts": {
            "$ref": "#/components/schemas/DQCounts"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          },
          "parser_version": {
            "title": "Parser Version",
            "type": "string"
          },
          "portfolio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio"
          },
          "publication_basis": {
            "default": "controlled_workflow",
            "title": "Publication Basis",
            "type": "string"
          },
          "publication_requires_override": {
            "title": "Publication Requires Override",
            "type": "boolean"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "publisher_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher Id"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "report_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Date"
          },
          "review_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Comment"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "source_sha256": {
            "title": "Source Sha256",
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "validating",
              "validated",
              "approved",
              "published",
              "failed",
              "rejected",
              "superseded",
              "withdrawn"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImportSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "uploader_id": {
            "title": "Uploader Id",
            "type": "string"
          },
          "validated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated At"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "id",
          "portfolio",
          "report_date",
          "version",
          "status",
          "duplicate",
          "source_sha256",
          "original_filename",
          "parser_version",
          "uploader_id",
          "reviewer_id",
          "publisher_id",
          "review_comment",
          "rejection_reason",
          "error_message",
          "created_at",
          "validated_at",
          "approved_at",
          "published_at",
          "snapshot_id",
          "summary",
          "dq_counts",
          "publication_requires_override"
        ],
        "title": "ImportRecord",
        "type": "object"
      },
      "ImportStatus": {
        "enum": [
          "draft",
          "validating",
          "validated",
          "approved",
          "published",
          "failed",
          "rejected",
          "superseded",
          "withdrawn"
        ],
        "title": "ImportStatus",
        "type": "string"
      },
...
      "ReferenceDictionaryStatus": {
        "additionalProperties": false,
        "properties": {
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "row_count",
          "updated_at"
        ],
        "title": "ReferenceDictionaryStatus",
        "type": "object"
      },
      "ReferenceDictionaryUploadResponse": {
        "additionalProperties": false,
        "properties": {
          "added_isins": {
            "items": {
              "type": "string"
            },
            "title": "Added Isins",
            "type": "array"
          },
          "changed_isins": {
            "items": {
              "type": "string"
            },
            "title": "Changed Isins",
            "type": "array"
          },
          "previous_row_count": {
            "title": "Previous Row Count",
            "type": "integer"
          },
          "removed_isins": {
            "items": {
              "type": "string"
            },
            "title": "Removed Isins",
            "type": "array"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          }
        },
        "required": [
          "row_count",
          "previous_row_count",
          "added_isins",
          "removed_isins",
          "changed_isins"
        ],
        "title": "ReferenceDictionaryUploadResponse",
        "type": "object"
      },
      "RejectionRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RejectionRequest",
        "type": "object"
      },
...
      "WithdrawalRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "WithdrawalRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OidcBearer": {
        "description": "Production OIDC access token. Development headers are local-only.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Portfolio Operations Insight API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/imports": {
      "get": {
        "operationId": "listImports",
        "parameters": [
          {
            "in": "query",
            "name": "portfolio",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Portfolio"
            }
          },
          {
            "in": "query",
            "name": "report_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Report Date"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ImportStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "include_withdrawn",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Withdrawn",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "List workbook imports",
        "tags": [
          "imports"
        ]
      },
      "post": {
        "operationId": "createImport",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRecord"
                }
              }
            },
            "description": "Identical source already exists"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Import an immutable OSIP workbook",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/export": {
      "get": {
        "description": "Download the reader-visible immutable import registry and its audit.",
        "operationId": "exportImportRegistry",
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Excel workbook with the visible import registry and audit"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Export visible import registry and audit as an Excel workbook",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{import_id}": {
      "get": {
        "operationId": "getImport",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read an import and its audit trail",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{import_id}/approve": {
      "post": {
        "operationId": "approveImport",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Independently approve a validated import",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{import_id}/comparison": {
      "get": {
        "operationId": "compareImport",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Compare an import with the prior approved version",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{import_id}/publish": {
      "post": {
        "operationId": "publishImport",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Publish an independently approved version",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{import_id}/reject": {
      "post": {
        "operationId": "rejectImport",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Reject an import without deleting evidence",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{import_id}/source": {
      "get": {
        "operationId": "downloadImportSource",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Download the exact original workbook",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{import_id}/withdraw": {
      "post": {
        "operationId": "withdrawImport",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Withdraw a published version without deleting its evidence",
        "tags": [
          "imports"
        ]
      }
    },
...
    "/api/v1/reference-data/classes-and-ratings": {
      "get": {
        "operationId": "getReferenceDictionaryStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceDictionaryStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read the current instrument classification/ratings dictionary status",
        "tags": [
          "imports"
        ]
      },
      "post": {
        "operationId": "uploadReferenceDictionary",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadReferenceDictionary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceDictionaryUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Replace the instrument classification/ratings dictionary",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/reference-data/dividends": {
      "get": {
        "operationId": "getDividendDataStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DividendDataStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read dividend reference-data freshness",
        "tags": [
          "imports"
        ]
      },
      "post": {
        "operationId": "uploadDividendData",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadDividendData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DividendDataUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Replace the Bloomberg dividend reference file",
        "tags": [
          "imports"
        ]
      }
    },
...
    "/api/v1/source-rows/{source_row_id}/preview": {
      "get": {
        "description": "Return a small, read-only view centered on an immutable source cell.\n\n``source_row_id`` identifies either an OSIP ``SourceRow`` (persisted\nimmutable evidence) or, for multi-source domains such as Risk, a\n``DatasetRecord`` - those never persist a row-level table, so this path\nre-opens the original workbook from blob storage instead.",
        "operationId": "previewSourceRow",
        "parameters": [
          {
            "in": "path",
            "name": "source_row_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Row Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cell",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 2,
              "title": "Cell",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcePreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read an exact source row centered on a workbook cell",
        "tags": [
          "imports"
        ]
      }
    },
...
  }
}